Skip empty items in slice parameters
Empty items in slice parameters make parsing fail for numeric, boolean
and duration types. Given that an empty value is not valid in any of
those cases, we can safely ignore them. As a consequence, this update
makes it possible to provide an entirely empty parameter, that would map
to an empty slice. Before, there was no chance to make an empty slice
other than not providing the parameter at all.

Closes VividCortex/artifacts#2387.